Fix the footer so it works with non react router projects. (#110)
This fixes the `<Footer />` component by ensuring the import
path for `<ExternalLink />` doesn't rely on `react-router`.
This enables the `<Footer />` to be used in projects that
don't include `react-router`.

The components still have to be imported in isolation (i.e. without
using the barrel file, re #103). We may accordingly come up
with a more pluggable mechanism and/or split `react-router` tied
functionality into a separate module in the future.