Ensure the footer and page share the same background color. (#109)
This modifies the `<Footer />` so that it sets the background
of the page. This fixes the issue where the footer appears
to "hang" on the page on larger screens without a lot of content.

I added an option that lets users disable this behavior, in
the event that it causes artifacts / issues or they'd like to
set the page background separately.