Fix #178; Fix the `ScrollToTopOnPageChange` component. (#180)
The component wasn't broken, but rather the demo was modified
such that the top-level `<App />` was no longer remounted
on page changes (which is good, for performance sake).

We could accordingly say the "fix" involves putting the component
at the right spot in the component tree, but that seems like a
tricky thing to force our users to sort out. Instead we've now
bound the component to `react-router`, and respond explicitly
to history changes.